comment revised
List. java

I am not sure how to make a good java doc comment. I have some 

comments on that:

- addPerson
Axel: Prompts the user to enter a new contact's information and 

then store that information in an array (should be arraylist, 

which is myList).


- searchLastName
Axel: Searches myList for a last name provided by the user, 
then returns the person object (should be boolean and call 

toString method to print the information of the person object) 

to which said last name belongs.


- isInt
 Axel: Checks to see if user input is an int, returns false 

(should be boolean) if it is not.